backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
pydantic-settings>=2.6.0
email-validator>=2.2.0
python-dotenv>=1.0.1
numpy>=2.1.0
faiss-cpu>=1.9.0
google-genai>=1.0.0
pytest>=8.3.0
